backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1,<5
python-multipart==0.0.12
redis==5.2.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv==1.0.1
aiofiles==24.1.0
requests==2.32.3
